Uninstall: Fix find errors (#1277)
* If these folders do not exist, find fails, test if they exist first

fix cache folder

logs

fix parsers rom library

rom library .reset

Revert "rom library .reset"

This reverts commit d3f76bb8c18c0b5223e83f3932c55b52b97d89cf.

ignore emulators folder

-p

cleaner name

fuzzy fuzzy

+++ functions/uninstallEmuFP.sh
@@ -29,10 +29,15 @@ uninstallEmuFP() {
     flatpak uninstall "$ID" -y --system
 
     shName=$(echo "$scriptname" | awk '{print tolower($0)}')
-    find "${launcherPath}/" "${romsPath}/emulators" "${romsPath}/desktop/remoteplay" "${romsPath}/desktop/generic-applications" -maxdepth 1 -type f \( -iname "$shName.sh" -o -iname "$shName-emu.sh" \) | \
-    while read -r f
-    do
-        echo "deleting $f"
-        rm -f "$f"
+    for romfolder in "${launcherPath}/" "${romsPath}/emulators" "${romsPath}/desktop/remoteplay" "${romsPath}/desktop/generic-applications"; do
+        if [ -d "$romfolder" ]; then
+            find "$romfolder" -maxdepth 1 -type f \( -iname "$shName.sh" -o -iname "$shName-emu.sh" \) | \
+            while read -r f; do
+                echo "deleting $f"
+                rm -f "$f"
+            done
+        else
+            echo "Skipping. $romfolder does not exist."
+        fi
     done
 }
